chore: move GTM code to separate file
