+++ gatsby-ssr.js
@@ -4,7 +4,8 @@ import React from 'react'
 import { Partytown } from '@builder.io/partytown/react'
 
 import Layout from './src/Layout'
-import AnalyticsHandler, { googleTagManager } from './src/sdk/analytics'
+import AnalyticsHandler from './src/sdk/analytics'
+import { googleTagManager } from './src/sdk/analytics/googleTagManager'
 import { validateCart } from './src/sdk/cart/validate'
 import ErrorBoundary from './src/sdk/error/ErrorBoundary'
 import TestProvider from './src/sdk/tests'
